+++ CHANGELOG.md
@@ -10,8 +10,12 @@
 ## [Unreleased][]
 ### Added
 ### Changed
+- Background color can now be set to null to use it as overlay.
+- Better Y axis tick management for auto mode.
 ### Removed
 ### Fixed
+- Y axis ticks were sometimes not properly set to respect the top margin.
+- Fixed some setting not being set in some cases while using the menu.
 
 ## [0.3.1] - 2023-08-24
 ### Added